+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+-   Arguments & options attributes for automatic validation.
 -   NoValidate attribute indicating that no automatic validation should be performed.
 
 ### Changed